build: disable S3 dependencies when `s3` feature is disabled
This also makes the DIMSE backend mandatory for now. The crate feature `dimse` was removed as the project didn't compile without it. In a future version, we'll revisit this to make the DIMSE backend optional as well. This allows for use cases where users only need the S3 backend for example.